+++ CHANGELOG.md
@@ -430,7 +430,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Deprecated `LightningLoggerBase.agg_and_log_metrics` in favor of `LightningLoggerBase.log_metrics` ([#11832](https://github.com/PyTorchLightning/pytorch-lightning/pull/11832))
 
 
-- Deprecated `pytorch_lightning.profiler.AbstractProfiler` in favor of `pytorch_lightning.profiler.Profiler` ([#FILLTHISIN](https://github.com/PyTorchLightning/pytorch-lightning/pull/FILLTHISIN))
+- Deprecated `pytorch_lightning.profiler.AbstractProfiler` in favor of `pytorch_lightning.profiler.BaseProfiler` ([#FILLTHISIN](https://github.com/PyTorchLightning/pytorch-lightning/pull/FILLTHISIN))
 
 
 ### Removed